# reference/openapi.yaml
openapi: 3.0.3
info:
  title: MAGNUS API
  description: |
    MAGNUS game-master API for managing campaigns, terminals, characters, and players.

    **Authentication:** JWT Bearer tokens (HS256, 24h expiry). Obtain via `POST /auth/login`.

    **Access control:**
    - Anonymous — public + active resources only
    - Player — public + active + assigned campaigns, own characters only
    - Admin — all resources

    **404 policy:** The API returns 404 for both "not found" and "access denied" to avoid leaking existence.
  version: 1.0.0
servers:
  - url: /
    description: Local server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ObjectId:
      type: string
      pattern: '^[a-f0-9]{24}$'
      example: '65f1a2b3c4d5e6f7a8b9c0d1'
...
    StateVarType:
      type: string
      enum: [boolean, number, enum, string]
...
    MutationOp:
      type: string
      enum: [set, increment, toggle]

    # ─── State ───────────────────────────────────────────────────────────────────
    StateEntry:
      type: object
      required: [type, value, default]
      properties:
        type:
          $ref: '#/components/schemas/StateVarType'
        value:
          description: Current runtime value
        default:
          description: Default value (used for reset)
        values:
          type: array
          items:
            type: string
          description: Allowed values (enum type only)
    StateMap:
      type: object
      additionalProperties:
        $ref: '#/components/schemas/StateEntry'
      example:
        omega_activated:
          type: boolean
          value: false
          default: false
    FlatState:
      type: object
      additionalProperties: {}
      example:
        omega_activated: false
        counter: 5
    StateMutation:
      type: object
      required: [key, op]
      properties:
        key:
          type: string
          pattern: '^(local|global)\..+'
          example: global.counter
        op:
          $ref: '#/components/schemas/MutationOp'
        value:
          description: Required for op=set
        by:
          type: number
          description: Required for op=increment

    # ─── User ────────────────────────────────────────────────────────────────────
...
    CampaignResponse:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/ObjectId'
        name:
          type: string
        isActive:
          type: boolean
        isPublic:
          type: boolean
        state:
          $ref: '#/components/schemas/StateMap'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        players:
          type: array
          items:
            $ref: '#/components/schemas/ObjectId'
          description: Player IDs — only populated for admin callers
    CreateCampaignDto:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        isActive:
          type: boolean
          default: false
        isPublic:
          type: boolean
          default: false
...
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
        message:
          type: string
        error:
          type: string
  responses:
    Unauthorized:
      description: Missing, invalid, or expired token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated but insufficient role
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found or caller cannot access it
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
paths:
...
  /campaigns:
    get:
      tags: [Campaigns]
      summary: List campaigns (actor-filtered)
      security:
        - {}
        - bearerAuth: []
      responses:
        '200':
          description: Campaign list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CampaignResponse'

    post:
      tags: [Campaigns]
      summary: Create campaign (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCampaignDto'
      responses:
        '201':
          description: Campaign created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /campaigns/{id}/state:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/ObjectId'
    get:
      tags: [Campaign State]
      summary: Get flat global state for campaign
      security:
        - {}
        - bearerAuth: []
      responses:
        '200':
          description: Flat state map
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlatState'
        '404':
          $ref: '#/components/responses/NotFound'
  /campaigns/{id}/state/mutate:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/ObjectId'
    post:
      tags: [Campaign State]
      summary: Apply mutations to global campaign state
      security:
        - {}
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mutations]
              properties:
                mutations:
                  type: array
                  items:
                    $ref: '#/components/schemas/StateMutation'
      responses:
        '200':
          description: Updated flat state
          content:
            application/json:
              schema:
                type: object
                properties:
                  state:
                    $ref: '#/components/schemas/FlatState'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /campaigns/{id}/state/reset:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/ObjectId'
    post:
      tags: [Campaign State]
      summary: Reset all campaign state and all terminal states (admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: State after reset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlatState'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /campaigns/{id}/state/{key}/reset:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/ObjectId'
      - name: key
        in: path
        required: true
        schema:
          type: string
    post:
      tags: [Campaign State]
      summary: Reset a single campaign state variable (admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: State after reset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlatState'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─── Campaign Configuration ───────────────────────────────────────────────────
...